refactor: Consolidate ITask and IUser interfaces into task.model.ts and user.model.ts, removing redundancy
