+++ project/back/src/routes/index.route.ts
@@ -1,0 +1,4 @@
+import authRoutes from './auth.routes';
+import taskRoutes from './task.routes';
+
+export { authRoutes, taskRoutes };
