+++ project/back/src/types/index.ts
@@ -1,6 +1,6 @@
 // Models
 export { IUser } from './models/user.model';
-export { ITask } from './models/task.model';
+// La interfaz ITask ahora se exporta desde models/task.model.ts
 
 // DTOs
 export { RegisterDTO, LoginDTO } from './dto/auth.dto';
@@ -12,3 +12,5 @@ export { AuthResponse } from './responses/auth.response';
 // Common
 export { JwtPayload } from './common/jwt.types';
 export { AuthRequest } from './common/request.types';
+
+
