+++ project/back/src/types/task.types.ts
@@ -1,13 +1,4 @@
-import { Document, Types } from 'mongoose';
-
-export interface ITask extends Document {
-  title: string;
-  description: string;
-  completed: boolean;
-  userId: Types.ObjectId;
-  createdAt: Date;
-  updatedAt: Date;
-}
+// Eliminada la definición redundante de ITask, ahora se encuentra en models/task.model.ts
 
 export interface CreateTaskDTO {
   title: string;
