+++ project/back/src/types/task/task.model.type.ts
@@ -1,4 +1,4 @@
-import { Document, Types } from 'mongoose';
+import { Types } from 'mongoose';
 
 export interface ITask extends Document {
   title: string;
